+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The changes documented here do not include those from the original repository.
 ## [Unreleased]
 
 ## 2021-11-12
+
+- Test: Added setBackgroundJob unit tests for Android (https://outsystemsrd.atlassian.net/browse/RMET-1191)
+
 - Implementation of the background job (https://outsystemsrd.atlassian.net/browse/RMET-1133)
 
 ## 2021-11-11
@@ -39,7 +42,7 @@ The changes documented here do not include those from the original repository.
 
 ## [Version 1.0.0]
 
-## 2021-10-04<>
+## 2021-10-04
 - Refactorings for Android (https://outsystemsrd.atlassian.net/browse/RMET-1066)
 
 ## 2021-09-21